refactor: convert function based view to class based
